+++ src/rules/maximumcommonedgesubgraph_ilp.rs
@@ -128,7 +128,7 @@ impl ReduceTo<ILP<bool>> for MaximumCommonEdgeSubgraph {
         }
 
         // Objective: maximize the number of preserved labelled arcs.
-        let objective: Vec<(usize, f64)> = (0..num_y).map(|seq| (y_idx(seq), 1.0)).collect();
+        let objective: Vec<(usize, i64)> = (0..num_y).map(|seq| (y_idx(seq), 1)).collect();
 
         let target = ILP::new(num_vars, constraints, objective, ObjectiveSense::Maximize)
             .map_err(Self::target_construction)?;
