+++ src/rules/maximumdomaticnumber_ilp.rs
@@ -109,7 +109,7 @@ impl ReduceTo<ILP<bool>> for MaximumDomaticNumber<SimpleGraph> {
         }
 
         // Objective: maximize Σ y_i
-        let objective: Vec<(usize, f64)> = (0..n).map(|i| (n * n + i, 1.0)).collect();
+        let objective: Vec<(usize, i64)> = (0..n).map(|i| (n * n + i, 1)).collect();
 
         let target = ILP::new(num_vars, constraints, objective, ObjectiveSense::Maximize)
             .map_err(Self::target_construction)?;
