+++ src/unit_tests/rules/ilp_i64_ilp_bool.rs
@@ -5,7 +5,7 @@ use crate::solvers::ILPSolver;
 fn integer_ilp(
     bounds: &[(i64, i64)],
     constraints: Vec<LinearConstraint>,
-    objective: Vec<(usize, f64)>,
+    objective: Vec<(usize, i64)>,
     sense: ObjectiveSense,
 ) -> ILP<i64> {
     ILP::with_variables(
@@ -20,7 +20,7 @@ fn integer_ilp(
     .unwrap()
 }
 
-fn solve_via_bool(source: &ILP<i64>) -> Option<(Vec<i64>, f64)> {
+fn solve_via_bool(source: &ILP<i64>) -> Option<(Vec<i64>, i64)> {
     let reduction = ReduceTo::<ILP<bool>>::reduce_to(source).expect("reduction should succeed");
     let witness = ILPSolver::new().solve(reduction.target_problem()).ok()?;
     let source_solution = reduction.extract_solution(&witness).unwrap();
@@ -36,12 +36,12 @@ fn test_ilp_i64_to_ilp_bool_closed_loop() {
             LinearConstraint::le(vec![(0, 1), (1, 1)], 5),
             LinearConstraint::le(vec![(0, 4), (1, 7)], 28),
         ],
-        vec![(0, -5.0), (1, -6.0)],
+        vec![(0, -5), (1, -6)],
         ObjectiveSense::Minimize,
     );
     let (solution, objective) = solve_via_bool(&source).unwrap();
     assert!(source.is_feasible(&solution).unwrap());
-    assert_eq!(objective, -27.0);
+    assert_eq!(objective, -27);
 }
 
 #[test]
@@ -49,12 +49,12 @@ fn test_ilp_i64_to_ilp_bool_maximize() {
     let source = integer_ilp(
         &[(0, 4), (0, 3)],
         vec![LinearConstraint::le(vec![(0, 1), (1, 1)], 6)],
-        vec![(0, 3.0), (1, 5.0)],
+        vec![(0, 3), (1, 5)],
         ObjectiveSense::Maximize,
     );
     let (solution, objective) = solve_via_bool(&source).unwrap();
     assert!(source.is_feasible(&solution).unwrap());
-    assert_eq!(objective, 24.0);
+    assert_eq!(objective, 24);
 }
 
 #[test]
@@ -71,7 +71,7 @@ fn test_ilp_i64_to_ilp_bool_target_structure() {
     let source = integer_ilp(
         &[(0, 5), (0, 5)],
         vec![LinearConstraint::le(vec![(0, 1), (1, 1)], 5)],
-        vec![(0, 1.0)],
+        vec![(0, 1)],
         ObjectiveSense::Maximize,
     );
     let reduction = ReduceTo::<ILP<bool>>::reduce_to(&source).unwrap();
@@ -86,8 +86,8 @@ fn test_ilp_i64_to_ilp_bool_target_structure() {
 
 #[test]
 fn test_ilp_i64_to_ilp_bool_single_variable() {
-    let source = integer_ilp(&[(0, 7)], vec![], vec![(0, 1.0)], ObjectiveSense::Maximize);
-    assert_eq!(solve_via_bool(&source).unwrap(), (vec![7], 7.0));
+    let source = integer_ilp(&[(0, 7)], vec![], vec![(0, 1)], ObjectiveSense::Maximize);
+    assert_eq!(solve_via_bool(&source).unwrap(), (vec![7], 7));
 }
 
 #[test]
@@ -95,12 +95,12 @@ fn test_ilp_i64_to_ilp_bool_equality_constraint() {
     let source = integer_ilp(
         &[(0, 3), (0, 3)],
         vec![LinearConstraint::eq(vec![(0, 1), (1, 1)], 4)],
-        vec![(0, 1.0)],
+        vec![(0, 1)],
         ObjectiveSense::Minimize,
     );
     let (solution, objective) = solve_via_bool(&source).unwrap();
     assert!(source.is_feasible(&solution).unwrap());
-    assert_eq!(objective, 1.0);
+    assert_eq!(objective, 1);
 }
 
 #[test]
@@ -112,10 +112,10 @@ fn test_ilp_i64_to_ilp_bool_ge_constraint() {
             LinearConstraint::ge(vec![(1, 1)], 1),
             LinearConstraint::le(vec![(0, 1), (1, 1)], 5),
         ],
-        vec![(0, 1.0), (1, 1.0)],
+        vec![(0, 1), (1, 1)],
         ObjectiveSense::Maximize,
     );
-    assert_eq!(solve_via_bool(&source).unwrap().1, 5.0);
+    assert_eq!(solve_via_bool(&source).unwrap().1, 5);
 }
 
 #[test]
@@ -138,15 +138,15 @@ fn test_ilp_i64_to_ilp_bool_variable_fixed_at_zero() {
     let source = integer_ilp(
         &[(0, 0), (0, 3)],
         vec![],
-        vec![(1, 1.0)],
+        vec![(1, 1)],
         ObjectiveSense::Maximize,
     );
-    assert_eq!(solve_via_bool(&source).unwrap(), (vec![0, 3], 3.0));
+    assert_eq!(solve_via_bool(&source).unwrap(), (vec![0, 3], 3));
 }
 
 #[test]
 fn test_ilp_i64_to_ilp_bool_power_of_two_bound() {
-    let source = integer_ilp(&[(0, 7)], vec![], vec![(0, 1.0)], ObjectiveSense::Maximize);
+    let source = integer_ilp(&[(0, 7)], vec![], vec![(0, 1)], ObjectiveSense::Maximize);
     let reduction = ReduceTo::<ILP<bool>>::reduce_to(&source).unwrap();
     assert_eq!(reduction.target_problem().num_vars(), 3);
 }
@@ -154,7 +154,7 @@ fn test_ilp_i64_to_ilp_bool_power_of_two_bound() {
 #[test]
 fn test_ilp_i64_to_ilp_bool_preserves_sense() {
     for sense in [ObjectiveSense::Minimize, ObjectiveSense::Maximize] {
-        let source = integer_ilp(&[(0, 3)], vec![], vec![(0, 1.0)], sense);
+        let source = integer_ilp(&[(0, 3)], vec![], vec![(0, 1)], sense);
         let reduction = ReduceTo::<ILP<bool>>::reduce_to(&source).unwrap();
         assert_eq!(reduction.target_problem().sense(), sense);
     }
